requirements.txt
# Core dependencies
streamlit>=1.31.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.2

# Document processing
PyPDF2>=3.0.0
pdfplumber>=0.10.2
python-docx>=1.0.0

# Audio processing
openai-whisper>=20231117
sounddevice>=0.4.6
scipy>=1.11.3
wavio>=0.0.7

# LLM integration
ollama>=0.1.5

# Utilities
python-dotenv>=1.0.0
pandas>=2.1.0